+++ CHANGELOG.md
@@ -11,6 +11,7 @@ breaking changes in the upcoming 3.x release. This release is scheduled for
 We are happy to announce the following GA libraries. Unless specifically noted,
 the APIs in these libraries are stable, and are ready for production use.
 
+- [Developer Connect API](/google/cloud/developerconnect/README.md)
 - [Public Certificate Authority API](/google/cloud/publicca/README.md)
 
 ## v2.24.0 - 2024-05